redirect to dashboard if user is logged in
