+++ src/app/app-routing.module.ts
@@ -1,12 +1,12 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { AnonymousComponent } from './anonymous/anonymous.component';
-import { AuthGuard } from './auth.guard';
+import { AuthGuard } from './guards/auth/auth.guard';
 import { DashboardComponent } from './dashboard/dashboard.component';
 
 const routes: Routes = [
-  { path: '', component: AnonymousComponent},
-  { path: 'dashboard', component: DashboardComponent, canActivate: [AuthGuard]},
+  { path: '', component: AnonymousComponent, canActivate: [AuthGuard], data: {requiresLogin: false, redirectTo: '/dashboard'}},
+  { path: 'dashboard', component: DashboardComponent, canActivate: [AuthGuard], data: {requiresLogin: true, redirectTo: '/'}},
   { path: '**', redirectTo: '' }
 ];
