+++ src/app/guards/auth/auth.guard.ts
@@ -2,7 +2,6 @@ import { Injectable } from '@angular/core';
 import { ActivatedRouteSnapshot, CanActivate, Router, RouterStateSnapshot, UrlTree } from '@angular/router';
 import { AuthConfig, OAuthService } from 'angular-oauth2-oidc';
 import { Observable } from 'rxjs';
-import { AppConfigService } from './services/app-config/app-config.service';
 
 @Injectable({
   providedIn: 'root'
@@ -14,12 +13,14 @@ export class AuthGuard implements CanActivate {
    canActivate(
     route: ActivatedRouteSnapshot,
     state: RouterStateSnapshot): Observable<boolean | UrlTree> | Promise<boolean | UrlTree> | boolean | UrlTree {
+      const requiresLogin = route.data['requiresLogin'] as boolean || false;
       const isLoggedIn = this.authService.hasValidIdToken();
-      if(isLoggedIn)
+      if(requiresLogin == isLoggedIn)
       {
         return true;
       }
-      this.router.navigate(['/']);
+      const redirectTo = route.data['redirectTo'] as string;
+      this.router.navigate([redirectTo]);
       return false;
   }
